+++ lib/rubocop/cop/rails/has_many_or_has_one_dependent.rb
@@ -5,8 +5,9 @@ module RuboCop
     module Rails
       # This cop looks for `has_many` or `has_one` associations that don't
       # specify a `:dependent` option.
+      #
       # It doesn't register an offense if `:through` or `dependent: nil`
-      # was specified.
+      # is specified, or if the model is read-only.
       #
       # @example
       #   # bad
@@ -21,6 +22,15 @@ module RuboCop
       #     has_one :avatar, dependent: :destroy
       #     has_many :articles, dependent: nil
       #     has_many :patients, through: :appointments
+      #   end
+      #
+      #   class User < ActiveRecord::Base
+      #     has_many :comments
+      #     has_one :avatar
+      #
+      #     def readonly?
+      #       true
+      #     end
       #   end
       class HasManyOrHasOneDependent < Base
         MSG = 'Specify a `:dependent` option.'
@@ -59,8 +69,14 @@ module RuboCop
             (args) ...)
         PATTERN
 
+        def_node_matcher :readonly?, <<~PATTERN
+          (def :readonly?
+            (args)
+            (true))
+        PATTERN
+
         def on_send(node)
-          return if active_resource?(node.parent)
+          return if active_resource?(node.parent) || readonly_model?(node)
           return if !association_without_options?(node) && valid_options?(association_with_options?(node))
           return if valid_options_in_with_options_block?(node)
 
@@ -68,6 +84,12 @@ module RuboCop
         end
 
         private
+
+        def readonly_model?(node)
+          return false unless (parent = node.parent)
+
+          parent.each_descendant(:def).any? { |def_node| readonly?(def_node) }
+        end
 
         def valid_options_in_with_options_block?(node)
           return true unless node.parent